Refactor names from articles to notes
